% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mongolmaps-package.R
\docType{package}
\name{mongolmaps-package}
\alias{mongolmaps}
\alias{mongolmaps-package}
\title{mongolmaps: Maps and Administrative Boundaries of Mongolia and Ulaanbaatar}
\description{
Ready-to-use 'sf' maps of Mongolia at every openly available administrative level: country, economic regions, aimags (provinces), soums and Ulaanbaatar districts, and Ulaanbaatar khoroos (subdistricts), with codes and names for every bag. Provides a crosswalk between English and Cyrillic names, common spelling variants, ISO 3166-2 codes, humanitarian P-codes and National Statistics Office codes; helpers to join statistics to maps; thematic layers (settlements, rivers, lakes, roads, railways, airports, elevation, land cover, population and protected areas); and quick 'ggplot2' and 'leaflet' maps. Boundaries derive from the Common Operational Dataset published by the National Statistics Office of Mongolia and the United Nations Office for the Coordination of Humanitarian Affairs \url{https://data.humdata.org/dataset/cod-ab-mng}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/temuulene/mongolmaps}
  \item \url{https://temuulene.github.io/mongolmaps/}
  \item Report bugs at \url{https://github.com/temuulene/mongolmaps/issues}
}

}
\author{
\strong{Maintainer}: Temuulen Enebish \email{temuulen@gmail.com} [copyright holder]

Authors:
\itemize{
  \item Temuulen Enebish \email{temuulen@gmail.com} [copyright holder]
}

}
\keyword{internal}
